backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
apscheduler
requests
scikit-learn
pandas
numpy
python-dotenv
